add support for external replication channel

+++ internal/mysql/data.go
@@ -45,9 +45,9 @@ type replicationSettings struct {
 	SourceHost     string `db:"SourceHost"`
 	SourceUser     string `db:"SourceUser"`
 	SourcePassword string `db:"SourcePassword"`
-	SourcePort	   int64 `db:"SourcePort"`
+	SourcePort	   int `db:"SourcePort"`
 	SourceSslCa    string `db:"SourceSslCa"`
-	SourceDelay    int64 `db:"SourceDelay"`
+	SourceDelay    int `db:"SourceDelay"`
 }
 
 // SlaveStatusStruct contains SHOW SLAVE STATUS response
